+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Fixed duplicate inline image rendering
 - Fixed post action blank conversation pane - [#271](https://github.com/hkdb/aerion/issues/271)
 - Fixed separate smtp creds persistence - [#270](https://github.com/hkdb/aerion/issues/270)
+- Fixed plaintext reply/fwd - [#285](https://github.com/hkdb/aerion/issues/285)
 - Bumped flatpak build to Gnome 50 runtime